==> 1_intro_R.Rmd <==
---
title: "Introduction to R"
---

This is an RMarkdown file. It mixes code and text. We can compile it into an HTML that contains everything. To compile it, click on “Knit”.

We can execute a chuck of code by `Ctrl+Shift+Enter` (Windows) or `Command+Shift+Enter` (macOS).

See a list of shortcuts [here](https://bookdown.org/yihui/rmarkdown-cookbook/rstudio-shortcuts.html).

## Download R and Rstudio

https://cran.r-project.org/bin/windows/base/ (Windows)

https://cran.r-project.org/bin/macosx/ (macOS)

https://www.rstudio.com/products/rstudio/download/


## R basics

Everything is stored as a variable. It's different from SPSS in this respect. We can see all the variables in the "Environment". To assign something to a variable, use `<-`
```{r}
result <- 1+1
```
To view a variable
```{r}
result
View(result)
```

|                    | R  | Python |
| ------------------ | -- | ------ |
| Addition           | +  | +      |
| Subtraction        | -  | -      |
| Multiplication     | *  | *      |
| Division           | /  | /      |
| Exponentiation     | ^  | **     |
| Variable assignment| <- | =      |

To check data type
```{r}
class(result)
```
|                    | R         | Python              |
| ------------------ | --------- | ------------------- |
| 200                | integer   | int                 |
| 200.15             | numeric   | float               |
| TRUE / FALSE       | logical   | bool (True / False) |
| "lit review"       | character | str                 |
| To check data type | `class()` | `type()`            |


Logical comparison:

|    <  less than
|    >  greater than
|    <= less than or equal to
|    >= greater than or equal to
|    == equal to
|    != not equal to

## Load packages
```{r}
library(haven) # read data in SPSS, SAS or Stata formats
library(dplyr) # data wrangling
library(ggplot2) # plots
```

The libraries are loaded successfully if there's no error message. Warnings are usually harmless. Pay attention to the conflicts.

## Import data
```{r}
soss <- read_sav("./SOSS81Data.sav")
```
```{r}
#View(soss)
soss
```

## Data wrangling

To select columns we need
```{r}
df <- soss %>% select(caseid, weight, driving18, driving19, driving20, driving21, driving22a, driving22b, driving22c, driving23a, driving23b, driving23c, driving16, driving17, CD1, age, agecat7, CD5a, races, rac3, educat4, laborforce, inc, X1, X1_t, driving01)
```
` %>% ` (`Ctrl+Shift+M` or `Command+Shift+M`) is called "pipe". It chains functions together by passing the left hand side of the operator to the first argument of the right hand side of the operator. In the above chain, `soss` is passed to `select()`, and then the result of that is saved to `df`.

```{r}
df %>% head()
```

To select rows
```{r}
df %>% filter(age <= 20)
```

To re-order the dataset
```{r}
df %>% arrange(inc) # by income (from less to more)
#df %>% arrange(desc(inc)) # from more to less
```

```{r}
#soss$CD1 %>% attr("labels")
unique(soss$CD1)
```
```{r}
df %>% count(CD1)
```

To create a new variable
```{r}
df %>% mutate(gender = as.factor(CD1))
```

```{r}
df %>% filter(age <= 20) %>% group_by(inc) %>% count()
#df$inc %>% attr("labels")
```

## Virtualization

There are several ways to draw plots in R. `base graphics` is built in R. `ggplot2` looks better. `plotly` is gaining popularity.

```{r}
# base graphics
plot(driving01 ~ age,data  = df) # hours of driving by age
```
```{r}
hist(df$driving01)
```

```{r}
ggplot(data = df) +
  geom_point(aes(x=age, y=driving01), alpha=0.4)
```

Notice the warnings. It is because we haven't clean the data properly. Let's ignore it for now.

```{r}
ggplot(data = df) +
  geom_jitter(aes(x=age, y=driving01), alpha=0.4)
```
```{r}
ggplot(data = df) +
  geom_jitter(aes(x=age, y=driving01, color=as.factor(CD1)), alpha=0.7)
```

```{r}
ggplot(df) +
  geom_histogram(aes(x=driving17)) #Interest in owning or leasing an AV
#unique(df$driving17)
```

## Clean the SOSS data

```{r}
#gender: 0=male, 1=female
df <- df %>% mutate(gender = as.factor(CD1))
df$gender[df$gender==3] <- NA
df$gender[df$gender==8] <- NA
df$gender[df$gender==9] <- NA
df$gender <- recode(df$gender,`1`= 0, `2`=1)

#age: continuous
df <- df %>% mutate(age=as.numeric(age))
#agec:1=18-29, 2=30-39, 3=40-49, 4=50-59, 5=60-69, 6=70-79, 7=80+
df <- df %>% mutate(agec=as.factor(agecat7))
```
## Respondents profile (unweighted)

```{r}
df %>% group_by(gender) %>% summarize(n=n()) %>% mutate(freq=round((n/sum(n)*100),2))
```

Or we can write a function
```{r}
profile <- function(data,variable){
  data %>% group_by({{variable}}) %>% summarize(n=n()) %>% mutate(freq=round((n/sum(n)*100),2))
  }
```
```{r}
profile(df, agec)
```

To exclude NA 
```{r}
df %>% filter(!is.na(agec)) %>%  group_by(agec) %>% summarize(n=n()) %>% mutate(freq=round((n/sum(n)*100),2))
```

## Respondents profile (weighted)

```{r}
library(survey)
```

Documentation: https://cran.r-project.org/web/packages/survey/survey.pdf
```{r}
w <- svydesign(ids = ~1, data= df, weights = df$weight)
```
```{r}
data.frame(prop.table(svytable(~agec, design = w)))%>% mutate(n=round(Freq*length(which(!is.na(df$agec)))), percentage=round((Freq*100),2))
```
